requirements.txt
streamlit>=1.32.0
langchain>=0.1.12
langchain-community>=0.0.28
langchain-core>=0.1.30
langchain-openai>=0.0.5
beautifulsoup4>=4.12.3
requests>=2.31.0
pandas>=2.2.1
python-dotenv>=1.0.1
openai>=1.55.3